server/requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# LangChain & Ecosystem
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-chroma
langchain-huggingface

# Vectorstore
chromadb

# Embeddings
sentence-transformers

# PDF Parsing and OCR
PyPDF  # backend for PyPDFLoader via langchain_community
pytesseract
pdf2image
reportlab

# Environment Variables
python-dotenv

# Typing & Utilities
pydantic
requests
google-api-core # For handling Google API exceptions like rate limits

# Logging (optional but recommended)
loguru

# File Uploads
aiofiles

python-multipart  # for handling file uploads in FastAPI
